diplom_3_3
