+++ pages/profile_page.py
@@ -5,7 +5,6 @@ import helper
 from data import UrlsData
 from pages.base_page import BasePage
 from locators.profile_locators import ProfileLocators as PL
-from locators.login_locators import LoginLocators as LL
 
 
 class ProfilePage(BasePage):
@@ -47,30 +46,3 @@ class ProfilePage(BasePage):
             return True
         else:
             return False
-
-    @allure.step('Авторизация')
-    def authorization(self):
-        name = helper.helper_name()
-        password = helper.helper_password()
-        email = helper.helper_email()
-
-        self.open_url(UrlsData.REGISTER)
-        self.find_element_visibility(LL.FIELD_NAME_FOR_REG)
-
-        self.set_text(LL.FIELD_NAME_FOR_REG, name)
-        self.set_text(LL.FIELD_EMAIL_FOR_REG, email)
-        self.set_text(LL.FIELD_PASSWORD_FOR_REG, password)
-
-        self.click_element_if_clickable(LL.BUTTON_FOR_REG)
-        self.find_element_not_visibility(LL.BUTTON_FOR_REG)
-
-        self.find_element_visibility(LL.FIELD_EMAIL_FOR_LOGIN)
-        self.set_text(LL.FIELD_EMAIL_FOR_LOGIN, email)
-        self.set_text(LL.FIELD_PASSWORD_FOR_LOGIN, password)
-
-        self.click_element_if_clickable(LL.BUTTON_FOR_LOGIN)
-        self.find_element_not_visibility(LL.BUTTON_FOR_LOGIN)
-
-    def precondition_for_tests(self):
-        self.authorization()
-        self.open_profile_page()
